Initial usage of external winkernel module
