+++ pkg/storecmds/storecmds_windows.go
@@ -1,0 +1,5 @@
+package storecmds
+
+import (
+	_ "sigs.k8s.io/kpng/backends/windows/kernelspace"
+)
